Split arena-specific functionality into separate classes :)
