+++ Assets/_MobileArcade/Common/Scripts/PlayerManager/PlayerManager.cs
@@ -12,30 +12,26 @@ public class PlayerManager : MonoBehaviour {
   // Distance from centerpoint
   [SerializeField] float dist;
 
-  List<Player> players;
+  [HideInInspector] public List<Player> players;
+
   Transform xform;
 
   void OnEnable () {
-	Physics.gravity = new Vector3(0, -0.5f, 0);
-
     xform = GetComponent<Transform>();
 
     players = new List<Player>();
 
     if (useWithoutPhone) {
       SpawnNewPlayer();
-      SpawnNewPlayer();  
+      SpawnNewPlayer();
     }
-
-	CircularSpawnPoints();
-	AssignPlayersColor();
   }
 
   void OnDisable () {
     for (int i = 0; i < players.Count; i++) {
 	  if (players[i] == null)
 	  	continue;
-			
+
       Destroy(players[i].gameObject);
     }
 
@@ -65,25 +61,4 @@ public class PlayerManager : MonoBehaviour {
 
     return player;
   }
-
-	void CircularSpawnPoints()
-	{
-		for (int i = 0; i < players.Count; i++)
-		{
-			float _x = Mathf.Cos((2*Mathf.PI / players.Count) * i);
-			float _z = Mathf.Sin((2*Mathf.PI / players.Count) * i);
-
-			players[i].transform.position = centerPoint + new Vector3(_x, 0, _z) * dist;
-		}
-	}
-
-	void AssignPlayersColor()
-	{
-		for (int i = 0; i < players.Count; i++)
-		{
-			HSBColor _hsbColor = new HSBColor((1 / (float) players.Count) * i, 1, 1, 1);
-			Color _color = HSBColor.ToColor(_hsbColor);
-			players[i].GetComponent<MeshRenderer>().material.color = _color;
-		}
-	}
 }
