+++ Assets/_MobileArcade/MinigameArena/Scripts.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 8e34fdbeacda85a46a0994ced3452906
+folderAsset: yes
+timeCreated: 1519767277
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
