+++ Assets/_MobileArcade/MinigameArena/Scripts/ArenaManager.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: cb5e3f236ba02a4478ca4962a71b43b5
+timeCreated: 1520183905
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
