+++ Assets/_MobileArcade/MinigameArena/Scripts/ArenaPlayer.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: f64412dd69ea3ed4b92fc23b7d292fc5
+timeCreated: 1520182639
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
